Switch to boinccmd
We are unable to pass arguments using the envar without encoutering errors. To circumvent this, running boinccmd instead of boinc seems to be the solution.
Logs:
The command-line options for /usr/bin/boinc are intended for debugging.
The recommended command-line interface is a separate program,'boinccmd'.
Run boinccmd in the same directory as /usr/bin/boinc.